requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.0.3

# RAG and LLM
langchain==0.0.352
langchain-community==0.0.16
langchain-core==0.1.12
langchain-text-splitters==0.0.1
qdrant-client==1.7.0
sentence-transformers==2.2.2
ollama==0.1.5

# Document Processing
unstructured==0.11.2
markdown==3.5.1
python-frontmatter==1.0.0

# Scheduler
apscheduler==3.10.4

# Frontend
streamlit==1.28.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
pylint==3.0.2
